+++ src/knockout/kosurvey.ts
@@ -7,6 +7,7 @@ import { LocalizableString } from "survey-core";
 import { ItemValue } from "survey-core";
 import { ImplementorBase } from "./kobase";
 import { getElement } from "survey-core";
+import { ILoadFromJSONOptions } from "survey-core";
 
 CustomWidgetCollection.Instance.onCustomWidgetAdded.add(customWidget => {
   if (customWidget.widgetJson.isDefaultRender) return;
@@ -126,9 +127,9 @@ export class Survey extends SurveyModel {
   public render(element: any = null): void {
     this.implementor.render(element);
   }
-  public fromJSON(json: any) {
+  public fromJSON(json: any, options?: ILoadFromJSONOptions): void {
     if (!json) return;
-    super.fromJSON(json);
+    super.fromJSON(json, options);
     this.locStrsChanged();
   }
   public getHtmlTemplate(): string {