+++ src/question_paneldynamic.ts
@@ -286,7 +286,7 @@ export class QuestionPanelDynamicModel extends Question
     this.registerPropertyChangedHandlers(["panelsState"], () => {
       this.setPanelsState();
     });
-    this.registerPropertyChangedHandlers(["isMobile"], () => {
+    this.registerPropertyChangedHandlers(["isMobile", "newPanelPosition", "showRangeInProgress", "renderMode"], () => {
       this.updateFooterActions();
     });
     this.registerPropertyChangedHandlers(["allowAddPanel"], () => { this.updateNoEntriesTextDefaultLoc(); });
@@ -911,6 +911,8 @@ export class QuestionPanelDynamicModel extends Question
    * Specifies whether users are allowed to add new panels.
    *
    * Default value: `true`
+   *
+   * By default, users add new panels to the end. If you want to let users insert a new panel after the current panel, set the [`newPanelPosition`](https://surveyjs.io/form-library/documentation/api-reference/dynamic-panel-model#newPanelPosition) property to `"next"`.
    * @see canAddPanel
    * @see allowRemovePanel
    */
@@ -919,6 +921,22 @@ export class QuestionPanelDynamicModel extends Question
   }
   public set allowAddPanel(val: boolean) {
     this.setPropertyValue("allowAddPanel", val);
+  }
+  /**
+   * Specifies the position of newly added panels.
+   *
+   * Possible values:
+   *
+   * - `"last"` (default) - New panels are added to the end.
+   * - `"next"` - New panels are inserted after the current panel.
+   * @see allowAddPanel
+   * @see addPanel
+   */
+  public get newPanelPosition(): string {
+    return this.getPropertyValue("newPanelPosition");
+  }
+  public set newPanelPosition(val: string) {
+    this.setPropertyValue("newPanelPosition", val);
   }
   /**
    * Specifies whether users are allowed to delete panels.
@@ -1002,7 +1020,6 @@ export class QuestionPanelDynamicModel extends Question
   }
   public set showRangeInProgress(val: boolean) {
     this.setPropertyValue("showRangeInProgress", val);
-    this.updateFooterActions();
     this.fireCallback(this.currentIndexChangedCallback);
   }
   /**
@@ -1021,7 +1038,6 @@ export class QuestionPanelDynamicModel extends Question
   }
   public set renderMode(val: string) {
     this.setPropertyValue("renderMode", val);
-    this.updateFooterActions();
     this.fireCallback(this.renderModeChangedCallback);
   }
   public get tabAlign(): "center" | "left" | "right" {
@@ -1089,7 +1105,8 @@ export class QuestionPanelDynamicModel extends Question
    */
   public get canAddPanel(): boolean {
     if (this.isDesignMode) return false;
-    if (this.isDefaultV2Theme && !this.legacyNavigation && !this.isRenderModeList && this.currentIndex < this.visiblePanelCount - 1) {
+    if (this.isDefaultV2Theme && !this.legacyNavigation && !this.isRenderModeList &&
+      (this.currentIndex < this.visiblePanelCount - 1 && this.newPanelPosition !== "next")) {
       return false;
     }
     return (
@@ -1229,6 +1246,8 @@ export class QuestionPanelDynamicModel extends Question
    * @param index *(Optional)* An index at which to insert the new panel. `undefined` adds the panel to the end or inserts it after the current panel if [`renderMode`](https://surveyjs.io/form-library/documentation/api-reference/dynamic-panel-model#renderMode) is `"tab"`. A negative index (for instance, -1) adds the panel to the end in all cases, regardless of the `renderMode` value.
    * @see panelCount
    * @see panels
+   * @see allowAddPanel
+   * @see newPanelPosition
    */
   public addPanel(index?: number): PanelModel {
     const curIndex = this.currentIndex;
@@ -1289,11 +1308,19 @@ export class QuestionPanelDynamicModel extends Question
    */
   public removePanelUI(value: any): void {
     if (!this.canRemovePanel) return;
-    if(this.confirmDelete) {
+    if(this.isRequireConfirmOnDelete(value)) {
       confirmActionAsync(this.confirmDeleteText, () => { this.removePanel(value); });
     } else {
       this.removePanel(value);
     }
+  }
+  public isRequireConfirmOnDelete(val: any): boolean {
+    if(!this.confirmDelete) return false;
+    const index = this.getVisualPanelIndex(val);
+    if(index < 0 || index >= this.visiblePanelCount) return false;
+    const panelValue = this.visiblePanels[index].getValue();
+    return !this.isValueEmpty(panelValue) &&
+      (this.isValueEmpty(this.defaultPanelValue) || !this.isTwoValueEquals(panelValue, this.defaultPanelValue));
   }
   /**
    * Switches Dynamic Panel to the next panel. Returns `true` in case of success, or `false` if `renderMode` is `"list"` or the current panel contains validation errors.
@@ -1521,6 +1548,7 @@ export class QuestionPanelDynamicModel extends Question
     if(this.isReadOnly || !this.allowAddPanel) {
       this.updateNoEntriesTextDefaultLoc();
     }
+    this.updateFooterActions();
     super.onSurveyLoad();
   }
   public onFirstRendering() {
@@ -2299,6 +2327,7 @@ Serializer.addClass(
     { name: "noEntriesText:text", serializationProperty: "locNoEntriesText" },
     { name: "allowAddPanel:boolean", default: true },
     { name: "allowRemovePanel:boolean", default: true },
+    { name: "newPanelPosition", choices: ["next", "last"], default: "last", category: "layout" },
     {
       name: "panelCount:number",
       isBindable: true,