+++ src/themes/doubleborder-dark-panelless.ts
@@ -28,6 +28,7 @@ const Theme = {
     "--sjs-shadow-medium": "0px 0px 0px 2px rgba(255, 255, 255, 0.1)",
     "--sjs-shadow-large": "0px 8px 16px 0px rgba(0, 0, 0, 0.1)",
     "--sjs-shadow-inner": "0px 0px 0px 2px rgba(255, 255, 255, 0.1)",
+    "--sjs-shadow-inner-reset": "0px 0px 0px 0px rgba(255, 255, 255, 0.1)",
     "--sjs-border-light": "rgba(255, 255, 255, 0.1)",
     "--sjs-border-default": "rgba(255, 255, 255, 0.1)",
     "--sjs-border-inside": "rgba(255, 255, 255, 0.08)",